Extract AuthGuard component from page.tsx
Refactored authentication logic into reusable wrapper component:

- Created AuthGuard component handling session loading and sign-in UI
- Extracted authentication flow and conditional rendering logic
